+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## UNRELEASED
 
+- Fix: Don't update last_sent to current time on every poll.
 - Feature: Allow for infinite retries in DB poller.
 
 # 1.22.2 - 2023-05-10