+++ CHANGELOG.md
@@ -7,9 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## UNRELEASED
 
+# 1.22.4 - 2023-07-05
+- Feature: Add support for message headers.
+
+# 1.22.3 - 2023-06-13
+
 - Fix: Don't update last_sent to current time on every poll.
 - Feature: Allow for infinite retries in DB poller.
-- Feature: Add support for message headers.
 
 # 1.22.2 - 2023-05-10
 - Feature: Add `DEIMOS_TASK_NAME` env variable when running a task (consumer, DB poller, DB producer).